+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 - Add the "Last Duration" field in the "My Checks" page (#257)
 - Add "last_duration" attribute to the Check API resource (#257)
 - Upgrade to psycopg2 2.8.3
+- Add Go usage example
 
 ### Bug Fixes
 - Usernames now are uuid3(const, email). Prevents multiple accts with same email (#290)