+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file.
 - Signup form sets the "auto-login" cookie to avoid an extra click during first login
 - Autofocus the email field in the signup form, and submit on enter key
 - Add support for OpsGenie EU region (#294)
+- Update OpsGenie logo and setup illustrations
 
 ### Bug Fixes
 - Prevent double-clicking the submit button in signup form